+++ apps/native-component-list/src/screens/ActionSheetScreen.tsx
@@ -1,16 +1,95 @@
 import { ActionSheetProvider, connectActionSheet } from '@expo/react-native-action-sheet';
 import React from 'react';
-import { ScrollView, StyleSheet, Text, View } from 'react-native';
+import {
+  ScrollView,
+  ActionSheetIOS,
+  TouchableOpacityProps,
+  TouchableOpacity,
+  StyleSheet,
+  Text,
+  View,
+  Platform,
+} from 'react-native';
 
 import ShowActionSheetButton from '../components/ShowActionSheetButton';
+import { Colors } from '../constants';
 
-const ActionSheetScreen = () => (
+const Button: React.FunctionComponent<TouchableOpacityProps> = props => {
+  return (
+    <TouchableOpacity
+      onPress={props.onPress}
+      style={{
+        paddingHorizontal: 15,
+        paddingVertical: 10,
+        borderRadius: 3,
+        backgroundColor: Colors.tintColor,
+        marginRight: 10,
+      }}>
+      <Text
+        style={{
+          color: '#fff',
+        }}>
+        {props.children}
+      </Text>
+    </TouchableOpacity>
+  );
+};
+
+const ActionSheetProviderScreen = () => (
   <ActionSheetProvider>
     {/*
     // @ts-ignore */}
     <App />
   </ActionSheetProvider>
 );
+
+function ActionSheetScreen() {
+  return (
+    <ScrollView>
+      <ActionSheetProviderScreen />
+      {Platform.OS === 'ios' && <ActionSheetIOSExample />}
+    </ScrollView>
+  );
+}
+
+function ActionSheetIOSExample() {
+  const showActionSheet = () => {
+    ActionSheetIOS.showActionSheetWithOptions(
+      {
+        options: ['Option 0', 'Option 1', 'Delete', 'Cancel'],
+        cancelButtonIndex: 3,
+        destructiveButtonIndex: 2,
+      },
+      buttonIndex => {
+        console.log({ buttonIndex });
+      }
+    );
+  };
+
+  const showShareSheet = () => {
+    ActionSheetIOS.showShareActionSheetWithOptions(
+      {
+        url: 'https://expo.io',
+        message: 'message to go with the shared url',
+        subject: 'a subject to go in the email heading',
+      },
+      error => alert(error),
+      (success, method) => {
+        if (success) {
+          alert(`Shared via ${method}`);
+        }
+      }
+    );
+  };
+
+  return (
+    <View style={{ flexDirection: 'row', padding: 10 }}>
+      <Button onPress={showActionSheet}>Action sheet</Button>
+
+      <Button onPress={showShareSheet}>Share sheet</Button>
+    </View>
+  );
+}
 
 ActionSheetScreen.navigationOptions = {
   title: 'Action Sheet',
